+++ test/Get-Instance.Tests.ps1
@@ -3,7 +3,7 @@
 Describe 'Get-Instance' {
 
     BeforeAll {
-        Import-Module $PSScriptRoot\..\Source\PsSqlTestServer.psd1 -Force -ErrorAction Stop
+        Import-Module $PSScriptRoot\..\src\PsSqlTestServer.psd1 -Force -ErrorAction Stop
     }
 
     It 'Provides a SQL server' {