+++ tasks/PsBuild.Tasks.ps1
@@ -1,8 +1,13 @@
+
+if ( -Not $PsBuildTaskBranch ) {
+    $PsBuildTaskBranch = 'main'
+}
+
 #region InvokeBuild
 
 task UpdateBuildTasks {
 	Invoke-WebRequest `
-		-Uri 'https://raw.githubusercontent.com/abbgrade/PsBuildTasks/main/Powershell/Build.Tasks.ps1' `
+		-Uri "https://raw.githubusercontent.com/abbgrade/PsBuildTasks/$PsBuildTaskBranch/Powershell/Build.Tasks.ps1" `
 		-OutFile "$PSScriptRoot\Build.Tasks.ps1"
 }
 
@@ -10,30 +15,44 @@ task UpdateBuildTasks {
 #region GitHub Actions
 
 task UpdateValidationWorkflow {
+    [System.IO.FileInfo] $file = "$PSScriptRoot/../.github/workflows/build-validation.yml"
+    New-Item -Type Directory $file.Directory -ErrorAction SilentlyContinue
     Invoke-WebRequest `
-        -Uri 'https://raw.githubusercontent.com/abbgrade/PsBuildTasks/main/GitHub/build-validation-matrix.yml' `
-        -OutFile "$PSScriptRoot\..\.github\workflows\build-validation.yml"
+        -Uri "https://raw.githubusercontent.com/abbgrade/PsBuildTasks/$PsBuildTaskBranch/GitHub/build-validation-matrix.yml" `
+        -OutFile $file
 }
 
 task UpdatePreReleaseWorkflow {
+    requires ModuleName
+    [System.IO.FileInfo] $file = "$PSScriptRoot\..\.github\workflows\pre-release.yml"
+    New-Item -Type Directory $file.Directory -ErrorAction SilentlyContinue
     Invoke-WebRequest `
-        -Uri 'https://raw.githubusercontent.com/abbgrade/PsBuildTasks/main/GitHub/pre-release-windows.yml' `
-        -OutFile "$PSScriptRoot\..\.github\workflows\pre-release.yml"
+        -Uri "https://raw.githubusercontent.com/abbgrade/PsBuildTasks/$PsBuildTaskBranch/GitHub/pre-release-windows.yml" |
+    ForEach-Object { $_ -replace 'MyModuleName', $ModuleName } |
+    Out-File $file -NoNewline
 }
 
 task UpdateReleaseWorkflow {
+    requires ModuleName
+    [System.IO.FileInfo] $file = "$PSScriptRoot\..\.github\workflows\release.yml"
+    New-Item -Type Directory $file.Directory -ErrorAction SilentlyContinue
     Invoke-WebRequest `
-        -Uri 'https://raw.githubusercontent.com/abbgrade/PsBuildTasks/main/GitHub/release-windows.yml' `
-        -OutFile "$PSScriptRoot\..\.github\workflows\release.yml"
+        -Uri "https://raw.githubusercontent.com/abbgrade/PsBuildTasks/$PsBuildTaskBranch/GitHub/release-windows.yml" |
+    ForEach-Object { $_ -replace 'MyModuleName', $ModuleName } |
+    Out-File $file -NoNewline
 }
 
 #endregion
 #region GitHub Pages
 
 task UpdateIndexPage {
+    New-Item -Type Directory "$PSScriptRoot\..\docs" -ErrorAction SilentlyContinue
     Invoke-WebRequest `
-        -Uri 'https://raw.githubusercontent.com/abbgrade/PsBuildTasks/main/docs/index.md' `
+        -Uri "https://raw.githubusercontent.com/abbgrade/PsBuildTasks/$PsBuildTaskBranch/docs/index.md" `
         -OutFile "$PSScriptRoot\..\docs\index.md"
+    Invoke-WebRequest `
+        -Uri "https://raw.githubusercontent.com/abbgrade/PsBuildTasks/$PsBuildTaskBranch/docs/_config.yml" `
+        -OutFile "$PSScriptRoot\..\docs\_config.yml"
 }
 
 #endregion
@@ -41,7 +60,7 @@ task UpdateIndexPage {
 
 task UpdateVsCodeTasks {
     Invoke-WebRequest `
-        -Uri 'https://raw.githubusercontent.com/abbgrade/PsBuildTasks/main/VsCode/tasks.json' `
+        -Uri "https://raw.githubusercontent.com/abbgrade/PsBuildTasks/$PsBuildTaskBranch/VsCode/tasks.json" `
         -OutFile "$PSScriptRoot\..\.vscode\tasks.json"
 }
 
@@ -50,7 +69,7 @@ task UpdateVsCodeTasks {
 
 task UpdatePsBuildTasksTasks {
     Invoke-WebRequest `
-        -Uri 'https://raw.githubusercontent.com/abbgrade/PsBuildTasks/main/tasks/PowerShell-Matrix.Tasks.ps1' `
+        -Uri "https://raw.githubusercontent.com/abbgrade/PsBuildTasks/$PsBuildTaskBranch/tasks/PowerShell-Matrix.Tasks.ps1" `
         -OutFile "$PSScriptRoot\PsBuild.Tasks.ps1"
 }