+++ CHANGELOG.md
@@ -7,9 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.3.0] - 2022-08-07
+
 ### Added
 
-- New-AzureInstance, Remove-AzureInstance
+- New-AzureInstance, Remove-AzureInstance and New-DatabaseName commands
 
 ### Changed